Implement SerializerFieldAdjuster
It will be the super class for serializers which need to remove a few of their fields.
